refactor(Card): move XDSCard to its own top-level module
Move XDSCard from Layout/Container/ to Card/ for better discoverability.
The component is now importable via '@xds/core/Card' with a dedicated
package export. The Layout re-export is preserved for backward compatibility.

Changes:
- Create packages/core/src/Card/ with component, index, and README
- Add ./Card export to package.json
- Add Card to top-level src/index.ts barrel export
- Update Layout/Container/XDSCard.tsx to re-export from Card/
- Update all storybook stories to import from '@xds/core/Card'
- Update Container README to reflect the move
