+++ packages/core/src/Card/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @file index.ts
+ * @input Imports XDSCard component
+ * @output Exports XDSCard component and types
+ * @position Entry point for @xds/core/Card module
+ *
+ * SYNC: When modified, update /packages/core/src/Card/README.md
+ */
+
+export {XDSCard} from './XDSCard';
+export type {XDSCardProps, SizeValue} from './XDSCard';
